+++ changelog.md
@@ -7,6 +7,7 @@ This version is not yet released.
 If you are reading this on the website, then these changes are live here.
 ### Language
 - [`join` `⊂`](https://uiua.org/docs/join) now always works with a scalar regardless of the shape of the other array
+- [`windows` `◫`](https://uiua.org/docs/windows) can now take negative window sizes
 
 ## 0.3.1 - 2023-11-20
 ### Interpreter